feat: add utils js-file with keyboard-config
