+++ index.html
@@ -1,11 +1,11 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="UTF-8">
-    <title>Virtual Keyboard</title>
-    <link rel="stylesheet" href="index.css">
+  <meta charset="UTF-8">
+  <title>Virtual Keyboard</title>
+  <link rel="stylesheet" href="index.css">
 </head>
 <body>
-<script src="index.js"></script>
+<script type="module" src="index.js"></script>
 </body>
 </html>
