+++ utils/constants.js
@@ -1,0 +1,652 @@
+const line_1 = {
+    'Backqoute': {
+        eng: {
+            lowerCase: '`',
+            upperCase: '~',
+        },
+        ru: {
+            lowerCase: 'ё',
+            upperCase: 'Ё',
+        },
+    },
+    'Digit1': {
+        eng: {
+            lowerCase: '1',
+            upperCase: '!',
+        },
+        ru: {
+            lowerCase: '1',
+            upperCase: '!',
+        },
+    },
+    'Digit2': {
+        eng: {
+            lowerCase: '2',
+            upperCase: '@',
+        },
+        ru: {
+            lowerCase: '2',
+            upperCase: '"',
+        },
+    },
+    'Digit3': {
+        eng: {
+            lowerCase: '3',
+            upperCase: '#',
+        },
+        ru: {
+            lowerCase: '3',
+            upperCase: '№',
+        },
+    },
+    'Digit4': {
+        eng: {
+            lowerCase: '4',
+            upperCase: '$',
+        },
+        ru: {
+            lowerCase: '4',
+            upperCase: ';',
+        },
+    },
+    'Digit5': {
+        eng: {
+            lowerCase: '5',
+            upperCase: '%',
+        },
+        ru: {
+            lowerCase: '5',
+            upperCase: '%',
+        },
+    },
+    'Digit6': {
+        eng: {
+            lowerCase: '6',
+            upperCase: '^',
+        },
+        ru: {
+            lowerCase: '6',
+            upperCase: ':',
+        },
+    },
+    'Digit7': {
+        eng: {
+            lowerCase: '7',
+            upperCase: '&',
+        },
+        ru: {
+            lowerCase: '7',
+            upperCase: '?',
+        },
+    },
+    'Digit8': {
+        eng: {
+            lowerCase: '8',
+            upperCase: '*',
+        },
+        ru: {
+            lowerCase: '8',
+            upperCase: '*',
+        },
+    },
+    'Digit9': {
+        eng: {
+            lowerCase: '9',
+            upperCase: '(',
+        },
+        ru: {
+            lowerCase: '9',
+            upperCase: '(',
+        },
+    },
+    'Digit0': {
+        eng: {
+            lowerCase: '0',
+            upperCase: ')',
+        },
+        ru: {
+            lowerCase: '0',
+            upperCase: ')',
+        },
+    },
+    'Minus': {
+        eng: {
+            lowerCase: '-',
+            upperCase: '_',
+        },
+        ru: {
+            lowerCase: '-',
+            upperCase: '_',
+        },
+    },
+    'Equal': {
+        eng: {
+            lowerCase: '=',
+            upperCase: '+',
+        },
+        ru: {
+            lowerCase: '=',
+            upperCase: '+',
+        },
+    },
+    'Backspace': {
+        eng: {
+            lowerCase: 'Backspace',
+            upperCase: 'Backspace',
+        },
+        ru: {
+            lowerCase: 'Backspace',
+            upperCase: 'Backspace',
+        },
+    }
+};
+const line_2 = {
+    'Tab': {
+        eng: {
+            lowerCase: 'Tab',
+            upperCase: 'Tab',
+        },
+        ru: {
+            lowerCase: 'Tab',
+            upperCase: 'Tab',
+        },
+    },
+    'KeyQ': {
+        eng: {
+            lowerCase: 'q',
+            upperCase: 'Q',
+        },
+        ru: {
+            lowerCase: 'й',
+            upperCase: 'Й',
+        },
+    },
+    'KeyW': {
+        eng: {
+            lowerCase: 'w',
+            upperCase: 'W',
+        },
+        ru: {
+            lowerCase: 'ц',
+            upperCase: 'Ц',
+        },
+    },
+    'KeyE': {
+        eng: {
+            lowerCase: 'e',
+            upperCase: 'E',
+        },
+        ru: {
+            lowerCase: 'у',
+            upperCase: 'У',
+        },
+    },
+    'KeyR': {
+        eng: {
+            lowerCase: 'r',
+            upperCase: 'R',
+        },
+        ru: {
+            lowerCase: 'к',
+            upperCase: 'К',
+        },
+    },
+    'KeyT': {
+        eng: {
+            lowerCase: 't',
+            upperCase: 'T',
+        },
+        ru: {
+            lowerCase: 'е',
+            upperCase: 'Е',
+        },
+    },
+    'KeyY': {
+        eng: {
+            lowerCase: 'y',
+            upperCase: 'Y',
+        },
+        ru: {
+            lowerCase: 'н',
+            upperCase: 'Н',
+        },
+    },
+    'KeyU': {
+        eng: {
+            lowerCase: 'u',
+            upperCase: 'U',
+        },
+        ru: {
+            lowerCase: 'г',
+            upperCase: 'Г',
+        },
+    },
+    'KeyI': {
+        eng: {
+            lowerCase: 'i',
+            upperCase: 'I',
+        },
+        ru: {
+            lowerCase: 'ш',
+            upperCase: 'Ш',
+        },
+    },
+    'KeyO': {
+        eng: {
+            lowerCase: 'o',
+            upperCase: 'O',
+        },
+        ru: {
+            lowerCase: 'щ',
+            upperCase: 'Щ',
+        },
+    },
+    'KeyP': {
+        eng: {
+            lowerCase: 'p',
+            upperCase: 'P',
+        },
+        ru: {
+            lowerCase: 'з',
+            upperCase: 'З',
+        },
+    },
+    'BracketLeft': {
+        eng: {
+            lowerCase: '[',
+            upperCase: '{',
+        },
+        ru: {
+            lowerCase: 'х',
+            upperCase: 'Х',
+        },
+    },
+    'BracketRight': {
+        eng: {
+            lowerCase: ']',
+            upperCase: '}',
+        },
+        ru: {
+            lowerCase: 'ъ',
+            upperCase: 'Ъ',
+        },
+    },
+    'Backslash': {
+        eng: {
+            lowerCase: '\\',
+            upperCase: '|',
+        },
+        ru: {
+            lowerCase: '\\',
+            upperCase: '/',
+        },
+    },
+    'Delete': {
+        eng: {
+            lowerCase: 'Del',
+            upperCase: 'Del',
+        },
+        ru: {
+            lowerCase: 'Del',
+            upperCase: 'Del',
+        },
+    },
+};
+const line_3 = {
+    'CapsLock': {
+        eng: {
+            lowerCase: 'CapsLock',
+            upperCase: 'CapsLock',
+        },
+        ru: {
+            lowerCase: 'CapsLock',
+            upperCase: 'CapsLock',
+        },
+    },
+    'KeyA': {
+        eng: {
+            lowerCase: 'a',
+            upperCase: 'A',
+        },
+        ru: {
+            lowerCase: 'ф',
+            upperCase: 'Ф',
+        },
+    },
+    'KeyS': {
+        eng: {
+            lowerCase: 's',
+            upperCase: 'S',
+        },
+        ru: {
+            lowerCase: 'ы',
+            upperCase: 'Ы',
+        },
+    },
+    'KeyD': {
+        eng: {
+            lowerCase: 'd',
+            upperCase: 'D',
+        },
+        ru: {
+            lowerCase: 'в',
+            upperCase: 'В',
+        },
+    },
+    'KeyF': {
+        eng: {
+            lowerCase: 'f',
+            upperCase: 'F',
+        },
+        ru: {
+            lowerCase: 'а',
+            upperCase: 'А',
+        },
+    },
+    'KeyG': {
+        eng: {
+            lowerCase: 'g',
+            upperCase: 'G',
+        },
+        ru: {
+            lowerCase: 'п',
+            upperCase: 'П',
+        },
+    },
+    'KeyH': {
+        eng: {
+            lowerCase: 'h',
+            upperCase: 'H',
+        },
+        ru: {
+            lowerCase: 'р',
+            upperCase: 'Р',
+        },
+    },
+    'KeyJ': {
+        eng: {
+            lowerCase: 'j',
+            upperCase: 'J',
+        },
+        ru: {
+            lowerCase: 'о',
+            upperCase: 'О',
+        },
+    },
+    'KeyK': {
+        eng: {
+            lowerCase: 'k',
+            upperCase: 'K',
+        },
+        ru: {
+            lowerCase: 'л',
+            upperCase: 'Л',
+        },
+    },
+    'KeyL': {
+        eng: {
+            lowerCase: 'l',
+            upperCase: 'L',
+        },
+        ru: {
+            lowerCase: 'д',
+            upperCase: 'Д',
+        },
+    },
+    'Semicolon': {
+        eng: {
+            lowerCase: ';',
+            upperCase: ':',
+        },
+        ru: {
+            lowerCase: 'ж',
+            upperCase: 'Ж',
+        },
+    },
+    'Quote': {
+        eng: {
+            lowerCase: '\'',
+            upperCase: '"',
+        },
+        ru: {
+            lowerCase: 'э',
+            upperCase: 'Э',
+        },
+    },
+    'Enter': {
+        eng: {
+            lowerCase: 'Enter',
+            upperCase: 'Enter',
+        },
+        ru: {
+            lowerCase: 'Enter',
+            upperCase: 'Enter',
+        },
+    },
+};
+const line_4 = {
+    'ShiftLeft': {
+        eng: {
+            lowerCase: 'Shift',
+            upperCase: 'Shift',
+        },
+        ru: {
+            lowerCase: 'Shift',
+            upperCase: 'Shift',
+        },
+    },
+    'KeyZ': {
+        eng: {
+            lowerCase: 'z',
+            upperCase: 'Z',
+        },
+        ru: {
+            lowerCase: 'я',
+            upperCase: 'Я',
+        },
+    },
+    'KeyX': {
+        eng: {
+            lowerCase: 'x',
+            upperCase: 'X',
+        },
+        ru: {
+            lowerCase: 'ч',
+            upperCase: 'Ч',
+        },
+    },
+    'KeyC': {
+        eng: {
+            lowerCase: 'c',
+            upperCase: 'C',
+        },
+        ru: {
+            lowerCase: 'с',
+            upperCase: 'С',
+        },
+    },
+    'KeyV': {
+        eng: {
+            lowerCase: 'v',
+            upperCase: 'V',
+        },
+        ru: {
+            lowerCase: 'м',
+            upperCase: 'М',
+        },
+    },
+    'KeyB': {
+        eng: {
+            lowerCase: 'b',
+            upperCase: 'B',
+        },
+        ru: {
+            lowerCase: 'и',
+            upperCase: 'И',
+        },
+    },
+    'KeyN': {
+        eng: {
+            lowerCase: 'n',
+            upperCase: 'N',
+        },
+        ru: {
+            lowerCase: 'т',
+            upperCase: 'Т',
+        },
+    },
+    'KeyM': {
+        eng: {
+            lowerCase: 'm',
+            upperCase: 'M',
+        },
+        ru: {
+            lowerCase: 'ь',
+            upperCase: 'Ь',
+        },
+    },
+    'Comma': {
+        eng: {
+            lowerCase: ',',
+            upperCase: '<',
+        },
+        ru: {
+            lowerCase: 'б',
+            upperCase: 'Б',
+        },
+    },
+    'Period': {
+        eng: {
+            lowerCase: '.',
+            upperCase: '>',
+        },
+        ru: {
+            lowerCase: 'ю',
+            upperCase: 'Ю',
+        },
+    },
+    'Slash': {
+        eng: {
+            lowerCase: '/',
+            upperCase: '?',
+        },
+        ru: {
+            lowerCase: '.',
+            upperCase: ',',
+        },
+    },
+    'ArrowUp': {
+        eng: {
+            lowerCase: '▲',
+            upperCase: '▲',
+        },
+        ru: {
+            lowerCase: '▲',
+            upperCase: '▲',
+        },
+    },
+    'ShiftRight': {
+        eng: {
+            lowerCase: 'Shift',
+            upperCase: 'Shift',
+        },
+        ru: {
+            lowerCase: 'Shift',
+            upperCase: 'Shift',
+        },
+    },
+};
+const line_5 = {
+    'ControlLeft': {
+        eng: {
+            lowerCase: 'Ctrl',
+            upperCase: 'Ctrl',
+        },
+        ru: {
+            lowerCase: 'Ctrl',
+            upperCase: 'Ctrl',
+        },
+    },
+    'MetaLeft': {
+        eng: {
+            lowerCase: 'Win',
+            upperCase: 'Win',
+        },
+        ru: {
+            lowerCase: 'Win',
+            upperCase: 'Win',
+        },
+    },
+    'AltLeft': {
+        eng: {
+            lowerCase: 'Alt',
+            upperCase: 'Alt',
+        },
+        ru: {
+            lowerCase: 'Alt',
+            upperCase: 'Alt',
+        },
+    },
+    'Space': {
+        eng: {
+            lowerCase: '',
+            upperCase: '',
+        },
+        ru: {
+            lowerCase: '',
+            upperCase: '',
+        },
+    },
+    'AltRight': {
+        eng: {
+            lowerCase: 'Alt',
+            upperCase: 'Alt',
+        },
+        ru: {
+            lowerCase: 'Alt',
+            upperCase: 'Alt',
+        },
+    },
+    'ArrowLeft': {
+        eng: {
+            lowerCase: '◄',
+            upperCase: '◄',
+        },
+        ru: {
+            lowerCase: '◄',
+            upperCase: '◄',
+        },
+    },
+    'ArrowDown': {
+        eng: {
+            lowerCase: '▼',
+            upperCase: '▼',
+        },
+        ru: {
+            lowerCase: '▼',
+            upperCase: '▼',
+        }
+    },
+    'ArrowRight': {
+        eng: {
+            lowerCase: '►',
+            upperCase: '►',
+        },
+        ru: {
+            lowerCase: '►',
+            upperCase: '►',
+        },
+    },
+    'ControlRight': {
+        eng: {
+            lowerCase: 'Ctrl',
+            upperCase: 'Ctrl',
+        },
+        ru: {
+            lowerCase: 'Ctrl',
+            upperCase: 'Ctrl',
+        },
+    },
+};
+
+export const linesArrow = [line_1, line_2, line_3, line_4, line_5];
